requirements.txt
django==5.0.3
psycopg2-binary==2.9.9
django-silk==5.1.0
django-ninja==1.2.0 
django-ninja-simple-jwt==0.5.0 